% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{as_field}
\alias{as_field}
\title{Add the field class to a data frame}
\usage{
as_field(data, row = NULL, col = NULL, env = NULL, ...)
}
\arguments{
\item{data}{A data frame to be converted to a field.}

\item{row}{A column name or expression to be used as the row identifier.}

\item{col}{A column name or expression to be used as the column identifier.}

\item{env}{A column name or expression to be used as the environment identifier.}

\item{...}{Not used.}
}
\description{
Specify the row, column, and environment identifiers to add the field class to a data frame.
}
