% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-checks.R
\name{cols_all_unique}
\alias{cols_all_unique}
\title{Identify columns with all unique values}
\usage{
cols_all_unique(data)
}
\arguments{
\item{data}{The data frame}
}
\value{
A character vector of column names with all unique values.
}
\description{
Identify columns with all unique values
}
