% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-checks.R
\name{cols_constant}
\alias{cols_constant}
\title{Identify constant columns}
\usage{
cols_constant(data, na.rm = FALSE)
}
\arguments{
\item{data}{The data frame}

\item{na.rm}{Should missing values be removed?}
}
\value{
A character vector of constant columns.
}
\description{
Similar to janitor::remove_constant, except this function
aims to identify the constant columns without removing them.
If there are no constant columns, it will return NA.
}
