% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-checks.R
\name{cols_identify_all}
\alias{cols_identify_all}
\title{Identify all columns with potential issues}
\usage{
cols_identify_all(data, cutoff = 1, na.rm = FALSE, print = TRUE)
}
\arguments{
\item{data}{The data frame}

\item{cutoff}{The minimum cutoff for the proportion of missing values.}

\item{na.rm}{Should missing values be removed?}

\item{print}{Print the output or not.}
}
\description{
Identify all columns with potential issues
}
