% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-checks.R
\name{cols_missing}
\alias{cols_missing}
\title{Identify columns that have certain proportion of missing values}
\usage{
cols_missing(data, cutoff = 1)
}
\arguments{
\item{data}{The data frame}

\item{cutoff}{The minimum cutoff for the proportion of missing values.}
}
\value{
A character vector of column names with missing values with at least a certain proportion.
}
\description{
Similar to janitor::remove_empty, except this function
aims to identify the columns without removing them.
If there are no identified columns, it will return NA.
}
