% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxe.R
\name{concurrence_matrix}
\alias{concurrence_matrix}
\alias{concurrence_table}
\title{Get the concurrence matrix or table}
\usage{
concurrence_matrix(data, gen = NULL, env = NULL)

concurrence_table(data, gen = NULL, env = NULL)
}
\arguments{
\item{data}{A data frame containing the MET data}

\item{gen}{The name of the genotype column}

\item{env}{The name of the environment column}
}
\value{
A concurrence matrix or table
}
\description{
Find the number of common genotypes across environments.
The diagonal shows the total number of genotypes at the corresponding environment.
The off-diagonal shows the number of common genotypes between the two environments.
The matrix is symmetric.
}
