% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropscan-package.R
\docType{package}
\name{cropscan-package}
\alias{cropscan}
\alias{cropscan-package}
\title{cropscan: Scanning Crop Field Experiments}
\description{
Initial data analysis for crop field experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emitanaka/cropscan}
  \item Report bugs at \url{https://github.com/emitanaka/cropscan/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

}
\keyword{internal}
