% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{field}
\alias{field}
\title{Construct a rectangular field}
\usage{
field(nrow = 1, ncol = 1, nenv = 1)
}
\arguments{
\item{nrow}{The number of rows.}

\item{ncol}{The number of columns.}

\item{nenv}{The number of envs.}
}
\description{
Construct a rectangular field
}
