% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\name{field_dim}
\alias{field_dim}
\title{Get the field dimensions}
\usage{
field_dim(data, trt = NULL, row = NULL, col = NULL, env = NULL)
}
\arguments{
\item{data}{A data frame or a field object.}

\item{trt}{A column names or expressions for treatment factors}

\item{row}{A column names or expressions for row identifiers}

\item{col}{A column names or expressions for column identifiers}

\item{env}{A column names or expressions for environment identifiers}
}
\description{
Get the field dimensions
}
\examples{
field_dim(met_pheno)
field_dim(field(12, 8, 3))
}
