% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-checks.R
\name{is_observational_unit}
\alias{is_observational_unit}
\title{Is it observational unit?}
\usage{
is_observational_unit(.data, ..., .message = TRUE)
}
\arguments{
\item{.data}{The data frame}

\item{...}{The columns in the data frame that potentially index the observational unit.}
}
\value{
A logical value.
}
\description{
Is it observational unit?
}
