% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{met_marker}
\alias{met_marker}
\title{Marker data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 323 rows and 17317 columns.
}
\usage{
met_marker
}
\description{
The marker information for the genotypes in the multi-environmental trial.
The marker data is coded as -1, 0, or 1.
}
\seealso{
Other met-data: 
\code{\link{met_pedigree}},
\code{\link{met_pheno}}
}
\concept{met-data}
\keyword{datasets}
