% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{met_pedigree}
\alias{met_pedigree}
\title{Pedigree data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9333 rows and 4 columns.
}
\usage{
met_pedigree
}
\description{
The pedigree information for the genotypes in the multi-environmental trial.
}
\seealso{
Other met-data: 
\code{\link{met_marker}},
\code{\link{met_pheno}}
}
\concept{met-data}
\keyword{datasets}
