% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{met_pheno}
\alias{met_pheno}
\title{Phenotypic data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9504 rows and 6 columns.
}
\usage{
met_pheno
}
\description{
A multi-environmental trial data with yield.
}
\seealso{
Other met-data: 
\code{\link{met_marker}},
\code{\link{met_pedigree}}
}
\concept{met-data}
\keyword{datasets}
