% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_field}
\alias{plot_field}
\title{Plot field}
\usage{
plot_field(
  data,
  fill = NULL,
  row = NULL,
  col = NULL,
  env = NULL,
  max_col = 40,
  max_env = 20,
  max_cat = 8
)
}
\arguments{
\item{data}{The data}

\item{fill}{The fill variable.}

\item{row, col}{The row and column name. If not supplied, it will try to detect one.}

\item{env}{The division of the env.}

\item{max_col}{The maximum number of columns to plot in a row if more than one env.}

\item{max_env}{The maximum number of envs.}

\item{max_cat}{The maximum number of categorical levels to plot (if fill is categorical).}
}
\description{
Plot field
}
