% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-checks.R
\name{rows_missing}
\alias{rows_missing}
\title{Identify rows that have certain proportion of missing values}
\usage{
rows_missing(data, cutoff = 1)
}
\arguments{
\item{data}{The data frame}

\item{cutoff}{The minimum cutoff for the proportion of missing values.}
}
\value{
An integer vector of rows with missing values with at least a certain proportion.
}
\description{
Similar to janitor::remove_empty, except this function
aims to identify the columns without removing them.
If there are no identified columns, it will return NA.
}
