% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_field}
\alias{plot_field}
\title{Plot field}
\usage{
plot_field(
  data,
  fill = NULL,
  row = NULL,
  col = NULL,
  trial = NULL,
  max_col = 40,
  max_trial = 20,
  max_cat = 8
)
}
\arguments{
\item{data}{The data}

\item{fill}{The fill variable.}

\item{row, col}{The row and column name. If not supplied, it will try to detect one.}

\item{trial}{The division of the trial.}

\item{max_col}{The maximum number of columns to plot in a row if more than one trial.}

\item{max_trial}{The maximum number of trials.}

\item{max_cat}{The maximum number of categorical levels to plot (if fill is categorical).}
}
\description{
Plot field
}
